requirements.txt
openai>=1.30.0
supabase>=2.4.0
langchain-text-splitters>=0.2.0
python-dotenv>=1.0.0
streamlit>=1.35.0
pandas>=2.0.0
openpyxl>=3.1.0